Improve detection of the INCLUDE syntax
A map in the YAML is only an INCLUDE syntax if it fits the INCLUDE
schema.

Also document the unordered nature of map, so in a merge include,
elements from the include file with the same keys with override the
original, regardless of where the elements may have appeared in the
original map.